problem-solving/data-structures/2d-array/2d-array.test.ts
import hourglassSum from './2d-array.ts';

describe('hourglassSum function', () => {
  it('should return the maximum hourglass sum (test case 1)', () => {
    const arr = [
      [1, 1, 1, 0, 0, 0],
      [0, 1, 0, 0, 0, 0],
      [1, 1, 1, 0, 0, 0],
      [0, 0, 2, 4, 4, 0],
      [0, 0, 0, 2, 0, 0],
      [0, 0, 1, 2, 4, 0],
    ];
    expect(hourglassSum(arr)).toBe(19);
  });

  it('should return the maximum hourglass sum (test case 2)', () => {
    const arr = [
      [1, 1, 1, 0, 0, 0],
      [0, 1, 0, 0, 0, 0],
      [1, 1, 1, 0, 0, 0],
      [0, 9, 2, -4, -4, 0],
      [0, 0, 0, -2, 0, 0],
      [0, 0, -1, -2, -4, 0],
    ];
    expect(hourglassSum(arr)).toBe(13);
  });

  it('should return the maximum hourglass sum (test case 3)', () => {
    const arr = [
      [-9, -9, -9, 1, 1, 1],
      [0, -9, 0, 4, 3, 2],
      [-9, -9, -9, 1, 2, 3],
      [0, 0, 8, 6, 6, 0],
      [0, 0, 0, -2, 0, 0],
      [0, 0, 1, 2, 4, 0],
    ];
    expect(hourglassSum(arr)).toBe(28);
  });

  it('should return the maximum hourglass sum (test case 4)', () => {
    const arr = [
      [1, 1, 1, 0, 0, 0],
      [0, 1, 0, 0, 0, 0],
      [1, 1, 1, 0, 0, 0],
      [0, 0, 2, 4, 4, 0],
      [0, 0, 0, 2, 0, 0],
      [0, 0, 1, 2, 4, 0],
    ];
    expect(hourglassSum(arr)).toBe(19);
  });
});
